Coding - Cast Graphic3d_NameOfTextureEnv for int comparison in ViewerTest_ViewerCommands

+++ src/Draw/TKViewerTest/ViewerTest/ViewerTest_ViewerCommands.cxx
@@ -8054,7 +8054,8 @@ static int VTextureEnv(Draw_Interpretor& /*theDI*/, int theArgNb, const char** t
       isOk = (!aTextureOpt.IsIntegerValue()
               || (aTextureOpt.IntegerValue() >= 0
                   && aTextureOpt.IntegerValue()
-                       < Graphic3d_NameOfTextureEnv::Graphic3d_NOT_ENV_UNKNOWN));
+                       < static_cast<int>(
+                         Graphic3d_NameOfTextureEnv::Graphic3d_NOT_ENV_UNKNOWN)));
 
       if (isOk && theArgNb == 11)
       {
